breaks main page into 2 and moves app to files-page, worked out sizing and search and routing.  starting to look like a config viewer, type search needs tweeking.
