+++ client/config-viewer/src/app/app.html
@@ -1,54 +1,29 @@
 ﻿<div class="app-container">
   <header class="app-header">
     <h1>{{ title() }}</h1>
-    
-    @if (selectedProject()) {
-      <div class="project-selector">
-        <label>Project:</label>
-        <select 
-          [value]="selectedProject()?.id"
-          (change)="selectProject(projects()[$any($event.target).selectedIndex])"
-        >
-          @for (project of projects(); track project.id) {
-            <option [value]="project.id">{{ project.name }}</option>
-          }
-        </select>
-      </div>
-    }
   </header>
 
-  <main class="app-main">
-    @if (loading()) {
-      <div class="loading">Loading projects...</div>
-    } @else if (error()) {
-      <div class="error">{{ error() }}</div>
-    } @else if (selectedProject()) {
-      <div class="split-pane">
-        <!-- Left: File Search -->
-        <div class="search-pane">
-          <app-searchable-list
-            [items]="files"
-            [loading]="searchLoading"
-            [columns]="fileColumns"
-            searchPlaceholder="Search files in {{ selectedProject()!.name }}..."
-            emptyMessage="Type to search for files"
-            noResultsMessage="No files found matching your search"
-            (searchChanged)="onSearchChanged($event)"
-            (itemSelected)="onFileSelected($event)"
-          />
-        </div>
+  <div class="app-body">
+    <!-- Left sidebar navigation -->
+    <nav class="app-nav">
+      <a routerLink="/files" routerLinkActive="active" class="nav-link">
+        <span class="nav-icon">📁</span>
+        <span class="nav-label">Files</span>
+      </a>
+      <a routerLink="/items" routerLinkActive="active" class="nav-link">
+        <span class="nav-icon">📊</span>
+        <span class="nav-label">Items</span>
+      </a>
+      <a routerLink="/projects" routerLinkActive="active" class="nav-link">
+        <span class="nav-icon">⚙️</span>
+        <span class="nav-label">Projects</span>
+      </a>
+    </nav>
+    
 
-        <!-- Right: File Viewer -->
-        <div class="viewer-pane">
-          <app-file-viewer
-            [fileName]="selectedFileName"
-            [filePath]="selectedFilePath"
-            [fileContent]="fileContent"
-            [loading]="fileLoading"
-            [error]="fileError"
-          />
-        </div>
-      </div>
-    }
-  </main>
+    <!-- Main content area -->
+    <main class="app-main"> 
+      <router-outlet />     
+    </main>
+  </div>
 </div>
