+++ client/config-viewer/src/app/features/files/files-page.component.scss
@@ -1,0 +1,205 @@
+.files-page {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  min-width: 0; // Critical: allows flex items to shrink below content size
+}
+
+.page-header {
+  padding: 12px 24px;
+  background: white;
+  border-bottom: 1px solid #dee2e6;
+  box-sizing: border-box; // Include padding in width calculation
+
+  .project-selector {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+
+    label {
+      font-size: 14px;
+      color: #495057;
+      font-weight: 600;
+    }
+
+    select {
+      padding: 6px 32px 6px 12px;
+      background: white;
+      border: 1px solid #ced4da;
+      border-radius: 4px;
+      font-size: 14px;
+      cursor: pointer;
+      min-width: 200px;
+    }
+  }
+}
+
+.content-area {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  width: 100%;
+  overflow: hidden;
+}
+
+.search-container {
+  position: relative;
+  width: calc(100vw - 200px);
+  padding: 16px 24px;
+  background: #f8f9fa;
+  border-bottom: 1px solid #dee2e6;
+  z-index: 10;
+  box-sizing: border-box; // Include padding in width calculation
+}
+
+.search-box {
+  position: relative;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  background: white;
+  border: 2px solid #ced4da;
+  border-radius: 6px;
+  transition: border-color 0.2s;
+
+  &:focus-within {
+    border-color: #0e639c;
+    box-shadow: 0 0 0 3px rgba(14, 99, 156, 0.1);
+  }
+
+  .search-icon {
+    padding-left: 12px;
+    font-size: 18px;
+    color: #6c757d;
+  }
+
+  .search-input {
+    flex: 1;
+    padding: 10px 12px;
+    border: none;
+    outline: none;
+    font-size: 14px;
+    background: transparent;
+  }
+
+  .clear-btn {
+    padding: 0 12px;
+    background: none;
+    border: none;
+    color: #6c757d;
+    font-size: 18px;
+    cursor: pointer;
+    transition: color 0.2s;
+
+    &:hover {
+      color: #dc3545;
+    }
+  }
+}
+
+.results-dropdown {
+  position: absolute;
+  top: calc(100% + 4px);
+  left: 24px;
+  right: 24px;
+  background: white;
+  border: 1px solid #dee2e6;
+  border-radius: 6px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  max-height: 800px;
+  overflow: auto;
+  display: flex;
+  flex-direction: column;
+
+  .results-info {
+    padding: 8px 16px;
+    font-size: 12px;
+    color: #6c757d;
+    background: #f8f9fa;
+    border-bottom: 1px solid #dee2e6;
+  }
+
+  .loading-state {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 12px;
+    padding: 32px;
+    color: #6c757d;
+
+    .spinner {
+      width: 20px;
+      height: 20px;
+      border: 2px solid #dee2e6;
+      border-top-color: #0e639c;
+      border-radius: 50%;
+      animation: spin 0.8s linear infinite;
+    }
+  }
+
+  .results-list {
+    flex: 1;
+    overflow-y: auto;
+  }
+
+  .result-item {
+    padding: 12px 16px;
+    cursor: pointer;
+    border-bottom: 1px solid #f8f9fa;
+    transition: background 0.15s;
+
+    &:hover {
+      background: #f8f9fa;
+    }
+
+    &.selected {
+      background: #e7f3ff;
+    }
+
+    .file-name {
+      font-size: 14px;
+      font-weight: 500;
+      color: #212529;
+      margin-bottom: 4px;
+    }
+
+    .file-path {
+      font-size: 12px;
+      color: #6c757d;
+      font-family: 'Consolas', 'Monaco', monospace;
+    }
+  }
+}
+
+.editor-area {
+  flex: 1;
+  display: flex;
+  padding: 0;
+  margin: 0;
+  overflow: hidden;
+  
+  app-file-viewer {
+    flex: 1;
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.loading, .error {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex: 1;
+  font-size: 16px;
+  color: #6c757d;
+}
+
+.error {
+  color: #dc3545;
+}
+
+@keyframes spin {
+  to { transform: rotate(360deg); }
+}
