+++ client/config-viewer/src/app/features/items/items-page.component.html
@@ -1,0 +1,24 @@
+<div class="items-page">
+  <div class="split-pane">
+    <!-- Left: Items Tree -->
+    <div class="tree-pane">
+      <div class="pane-header">
+        <h2>Items Hierarchy</h2>
+        <button class="btn-primary">+ New Item</button>
+      </div>
+      <div class="tree-content">
+        <p class="placeholder">Items tree will go here...</p>
+      </div>
+    </div>
+
+    <!-- Right: Item Details -->
+    <div class="details-pane">
+      <div class="pane-header">
+        <h2>Item Details</h2>
+      </div>
+      <div class="details-content">
+        <p class="placeholder">Select an item to view details...</p>
+      </div>
+    </div>
+  </div>
+</div>
