+++ client/config-viewer/src/app/features/items/items-page.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-items-page',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './items-page.component.html',
+  styleUrl: './items-page.component.scss',
+})
+export class ItemsPageComponent {
+  // TODO: Implement items tree navigation and details display
+}
